Feature: Refactoring

+++ src/main/java/org/imures/cashregister/catalog/repository/CatalogRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CatalogRepository extends JpaRepository<CatalogEntity, Integer>
+public interface CatalogRepository extends JpaRepository<CatalogEntity, Long>
 {
     CatalogEntity findByCatalogName(String name);
